requirements.txt
absl-py==1.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrdict==2.0.1
attrs==25.3.0
boto3==1.38.7
botocore==1.38.7
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
dill==0.4.0
dm-tree==0.1.9
docker==6.1.3
docker-pycreds==0.4.0
fastapi==0.115.12
fastapi-cli==0.0.1
fedml==0.9.6
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
gast==0.6.0
gensim==4.3.3
gevent==25.4.2
geventhttpclient==2.0.9
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
GPUtil==1.4.0
graphviz==0.8.4
greenlet==3.2.1
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib-metadata==1.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.18
namex==0.0.9
networkx==2.8.8
ntplib==0.4.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnx==1.17.0
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
paho-mqtt==1.6.1
pandas==2.2.3
pathtools==0.1.2
pillow==11.2.1
pluggy==1.5.0
prettytable==3.16.0
promise==2.3
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
py-machineid==0.7.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-rapidjson==1.20
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
requests==2.31.0
rich==14.0.0
s3transfer==0.12.0
scikit-learn==1.6.1
scipy==1.13.1
sentry-sdk==2.27.0
setproctitle==1.3.6
shortuuid==1.0.13
six==1.17.0
sklearn-compat==0.1.3
smart-open==6.3.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-model-optimization==0.8.0
termcolor==3.1.0
threadpoolctl==3.6.0
toposort==1.10
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
tritonclient==2.56.0
typer==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
wandb==0.13.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
